pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fintech-autogen"
version = "0.1.0"
description = "A fintech application leveraging AutoGen for portfolio management, market analysis, and financial education"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "autogen-agentchat>=0.5.7",
    "autogen-core>=0.5.7",
    "autogen-ext[magentic-one,openai]>=0.5.7",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "plotly>=5.3.0",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-openai>=0.0.1",
    "langchain-huggingface>=0.0.1",  # Added for HuggingFace embeddings
    "openai>=1.0.0",
    "yfinance>=0.2.0",
    "streamlit>=1.10.0",
    "python-dotenv>=0.19.0",
    "langchain-chroma>=0.1.0",
    "markitdown",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/fintech-autogen"
"Bug Tracker" = "https://github.com/yourusername/fintech-autogen/issues"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]